cleanup druid codebase
