+++ indexing-hadoop/src/test/java/io/druid/indexer/DeterminePartitionsJobTest.java
@@ -263,10 +263,6 @@ public class DeterminePartitionsJobTest
                 false,
                 null,
                 false,
-                false,
-                false,
-                null,
-                null,
                 false
             )
         )
