+++ indexing-hadoop/src/test/java/io/druid/indexer/JobHelperTest.java
@@ -112,10 +112,6 @@ public class JobHelperTest
                     "THISISMYACCESSKEY"
                 ),
                 false,
-                false,
-                false,
-                null,
-                null,
                 false
             )
         )
