+++ indexing-hadoop/src/test/java/io/druid/indexer/updater/HadoopConverterJobTest.java
@@ -200,10 +200,6 @@ public class HadoopConverterJobTest
                 false,
                 null,
                 false,
-                false,
-                false,
-                null,
-                null,
                 false
             )
         )
