+++ indexing-service/src/main/java/io/druid/indexing/common/task/IndexTask.java
@@ -132,11 +132,7 @@ public class IndexTask extends AbstractFixedIntervalTask
         null,
         null,
         shardSpec,
-        indexSpec,
-        null,
-        null,
-        null,
-        null
+        indexSpec
     );
   }
