+++ indexing-service/src/main/java/io/druid/indexing/common/task/RealtimeIndexTask.java
@@ -277,7 +277,6 @@ public class RealtimeIndexTask extends AbstractTask
         toolbox.getNewSegmentServerView(),
         toolbox.getQueryExecutorService(),
         toolbox.getIndexMerger(),
-        toolbox.getIndexMaker(),
         toolbox.getIndexIO()
     );
