+++ indexing-service/src/test/java/io/druid/indexing/common/task/IndexTaskTest.java
@@ -36,7 +36,6 @@ import io.druid.indexing.common.actions.TaskActionClientFactory;
 import io.druid.query.aggregation.AggregatorFactory;
 import io.druid.query.aggregation.LongSumAggregatorFactory;
 import io.druid.segment.IndexIO;
-import io.druid.segment.IndexMaker;
 import io.druid.segment.IndexMerger;
 import io.druid.segment.IndexSpec;
 import io.druid.segment.indexing.DataSchema;
@@ -69,7 +68,6 @@ public class IndexTaskTest
   private final IndexSpec indexSpec;
   private final ObjectMapper jsonMapper;
   private IndexMerger indexMerger;
-  private IndexMaker indexMaker;
   private IndexIO indexIO;
 
   public IndexTaskTest()
@@ -78,7 +76,6 @@ public class IndexTaskTest
     TestUtils testUtils = new TestUtils();
     jsonMapper = testUtils.getTestObjectMapper();
     indexMerger = testUtils.getTestIndexMerger();
-    indexMaker = testUtils.getTestIndexMaker();
     indexIO = testUtils.getTestIndexIO();
   }
 
@@ -260,7 +257,7 @@ public class IndexTaskTest
             return segment;
           }
         }, null, null, null, null, null, null, null, null, null, null, temporaryFolder.newFolder(),
-            indexMerger, indexMaker, indexIO
+            indexMerger, indexIO
         )
     );
