+++ indexing-service/src/test/java/io/druid/indexing/firehose/IngestSegmentFirehoseFactoryTimelineTest.java
@@ -54,7 +54,6 @@ import io.druid.query.aggregation.AggregatorFactory;
 import io.druid.query.aggregation.LongSumAggregatorFactory;
 import io.druid.query.filter.NoopDimFilter;
 import io.druid.segment.IndexIO;
-import io.druid.segment.IndexMaker;
 import io.druid.segment.IndexMerger;
 import io.druid.segment.IndexSpec;
 import io.druid.segment.incremental.IncrementalIndexSchema;
@@ -109,14 +108,12 @@ public class IngestSegmentFirehoseFactoryTimelineTest
 
   private static final ObjectMapper MAPPER;
   private static final IndexMerger INDEX_MERGER;
-  private static final IndexMaker INDEX_MAKER;
   private static final IndexIO INDEX_IO;
 
   static {
     TestUtils testUtils = new TestUtils();
     MAPPER = IngestSegmentFirehoseFactoryTest.setupInjectablesInObjectMapper(testUtils.getTestObjectMapper());
     INDEX_MERGER = testUtils.getTestIndexMerger();
-    INDEX_MAKER = testUtils.getTestIndexMaker();
     INDEX_IO = testUtils.getTestIndexIO();
   }
 
@@ -331,7 +328,6 @@ public class IngestSegmentFirehoseFactoryTimelineTest
           ),
           MAPPER,
           INDEX_MERGER,
-          INDEX_MAKER,
           INDEX_IO
       );
       final Injector injector = Guice.createInjector(
