+++ indexing-service/src/test/java/io/druid/indexing/worker/WorkerTaskMonitorTest.java
@@ -35,7 +35,6 @@ import io.druid.indexing.overlord.TestRemoteTaskRunnerConfig;
 import io.druid.indexing.overlord.ThreadPoolTaskRunner;
 import io.druid.indexing.worker.config.WorkerConfig;
 import io.druid.segment.IndexIO;
-import io.druid.segment.IndexMaker;
 import io.druid.segment.IndexMerger;
 import io.druid.segment.loading.SegmentLoaderConfig;
 import io.druid.segment.loading.SegmentLoaderLocalCacheManager;
@@ -73,7 +72,6 @@ public class WorkerTaskMonitorTest
   private Worker worker;
   private ObjectMapper jsonMapper;
   private IndexMerger indexMerger;
-  private IndexMaker indexMaker;
   private IndexIO indexIO;
 
   public WorkerTaskMonitorTest()
@@ -81,7 +79,6 @@ public class WorkerTaskMonitorTest
     TestUtils testUtils = new TestUtils();
     jsonMapper = testUtils.getTestObjectMapper();
     indexMerger = testUtils.getTestIndexMerger();
-    indexMaker = testUtils.getTestIndexMaker();
     indexIO = testUtils.getTestIndexIO();
   }
 
@@ -159,7 +156,6 @@ public class WorkerTaskMonitorTest
             ),
                 jsonMapper,
                 indexMerger,
-                indexMaker,
                 indexIO
             ),
             null
