+++ server/src/main/java/io/druid/segment/realtime/plumber/FlushingPlumberSchool.java
@@ -25,7 +25,6 @@ import com.metamx.emitter.service.ServiceEmitter;
 import io.druid.guice.annotations.Processing;
 import io.druid.query.QueryRunnerFactoryConglomerate;
 import io.druid.segment.IndexIO;
-import io.druid.segment.IndexMaker;
 import io.druid.segment.IndexMerger;
 import io.druid.segment.indexing.DataSchema;
 import io.druid.segment.indexing.RealtimeTuningConfig;
@@ -50,7 +49,6 @@ public class FlushingPlumberSchool extends RealtimePlumberSchool
   private final DataSegmentAnnouncer segmentAnnouncer;
   private final ExecutorService queryExecutorService;
   private final IndexMerger indexMerger;
-  private final IndexMaker indexMaker;
   private final IndexIO indexIO;
 
   @JsonCreator
@@ -61,7 +59,6 @@ public class FlushingPlumberSchool extends RealtimePlumberSchool
       @JacksonInject DataSegmentAnnouncer segmentAnnouncer,
       @JacksonInject @Processing ExecutorService queryExecutorService,
       @JacksonInject IndexMerger indexMerger,
-      @JacksonInject IndexMaker indexMaker,
       @JacksonInject IndexIO indexIO
   )
   {
@@ -74,7 +71,6 @@ public class FlushingPlumberSchool extends RealtimePlumberSchool
         null,
         queryExecutorService,
         indexMerger,
-        indexMaker,
         indexIO
     );
 
@@ -84,7 +80,6 @@ public class FlushingPlumberSchool extends RealtimePlumberSchool
     this.segmentAnnouncer = segmentAnnouncer;
     this.queryExecutorService = queryExecutorService;
     this.indexMerger = Preconditions.checkNotNull(indexMerger, "Null IndexMerger");
-    this.indexMaker = Preconditions.checkNotNull(indexMaker, "Null IndexMaker");
     this.indexIO = Preconditions.checkNotNull(indexIO, "Null IndexIO");
   }
 
@@ -107,7 +102,6 @@ public class FlushingPlumberSchool extends RealtimePlumberSchool
         segmentAnnouncer,
         queryExecutorService,
         indexMerger,
-        indexMaker,
         indexIO
     );
   }
