--- server/src/main/java/io/druid/server/bridge/BridgeCuratorConfig.java
@@ -1,29 +1,0 @@
-/*
- * Druid - a distributed column store.
- * Copyright 2012 - 2015 Metamarkets Group Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.druid.server.bridge;
-
-import io.druid.curator.CuratorConfig;
-import org.skife.config.Config;
-
-/**
- */
-public abstract class BridgeCuratorConfig extends CuratorConfig
-{
-  @Config("druid.bridge.zk.service.host")
-  public abstract String getParentZkHosts();
-}
