+++ server/src/main/java/io/druid/server/coordination/ZkCoordinator.java
@@ -1,30 +1,33 @@
 /*
- * Druid - a distributed column store.
- * Copyright 2012 - 2015 Metamarkets Group Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+* Licensed to Metamarkets Group Inc. (Metamarkets) under one
+* or more contributor license agreements. See the NOTICE file
+* distributed with this work for additional information
+* regarding copyright ownership. Metamarkets licenses this file
+* to you under the Apache License, Version 2.0 (the
+* "License"); you may not use this file except in compliance
+* with the License. You may obtain a copy of the License at
+*
+* http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing,
+* software distributed under the License is distributed on an
+* "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+* KIND, either express or implied. See the License for the
+* specific language governing permissions and limitations
+* under the License.
+*/
 
 package io.druid.server.coordination;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.base.Throwables;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Queues;
 import com.google.common.util.concurrent.SettableFuture;
-import com.google.common.util.concurrent.ThreadFactoryBuilder;
-import com.google.inject.Inject;
 import com.metamx.common.ISE;
 import com.metamx.common.concurrent.ScheduledExecutorFactory;
+import com.metamx.common.lifecycle.LifecycleStart;
+import com.metamx.common.lifecycle.LifecycleStop;
 import com.metamx.emitter.EmittingLogger;
 import io.druid.concurrent.Execs;
 import io.druid.segment.loading.SegmentLoaderConfig;
@@ -32,6 +35,11 @@ import io.druid.segment.loading.SegmentLoadingException;
 import io.druid.server.initialization.ZkPathsConfig;
 import io.druid.timeline.DataSegment;
 import org.apache.curator.framework.CuratorFramework;
+import org.apache.curator.framework.recipes.cache.ChildData;
+import org.apache.curator.framework.recipes.cache.PathChildrenCache;
+import org.apache.curator.framework.recipes.cache.PathChildrenCacheEvent;
+import org.apache.curator.framework.recipes.cache.PathChildrenCacheListener;
+import org.apache.curator.utils.ZKPaths;
 
 import java.io.File;
 import java.io.IOException;
@@ -41,7 +49,6 @@ import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledFuture;
@@ -50,17 +57,25 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  */
-public class ZkCoordinator extends BaseZkCoordinator
+public class ZkCoordinator implements DataSegmentChangeHandler
 {
   private static final EmittingLogger log = new EmittingLogger(ZkCoordinator.class);
 
+  private final Object lock = new Object();
+
   private final ObjectMapper jsonMapper;
+  private final ZkPathsConfig zkPaths;
   private final SegmentLoaderConfig config;
+  private final DruidServerMetadata me;
+  private final CuratorFramework curator;
   private final DataSegmentAnnouncer announcer;
   private final ServerManager serverManager;
   private final ScheduledExecutorService exec;
 
-  @Inject
+
+  private volatile PathChildrenCache loadQueueCache;
+  private volatile boolean started = false;
+
   public ZkCoordinator(
       ObjectMapper jsonMapper,
       SegmentLoaderConfig config,
@@ -72,17 +87,155 @@ public class ZkCoordinator extends BaseZkCoordinator
       ScheduledExecutorFactory factory
   )
   {
-    super(jsonMapper, zkPaths, config, me, curator);
-
     this.jsonMapper = jsonMapper;
+    this.zkPaths = zkPaths;
     this.config = config;
+    this.me = me;
+    this.curator = curator;
     this.announcer = announcer;
     this.serverManager = serverManager;
 
     this.exec = factory.create(1, "ZkCoordinator-Exec--%d");
   }
 
-  @Override
+  @LifecycleStart
+  public void start() throws IOException
+  {
+    synchronized (lock) {
+      if (started) {
+        return;
+      }
+
+      log.info("Starting zkCoordinator for server[%s]", me.getName());
+
+      final String loadQueueLocation = ZKPaths.makePath(zkPaths.getLoadQueuePath(), me.getName());
+      final String servedSegmentsLocation = ZKPaths.makePath(zkPaths.getServedSegmentsPath(), me.getName());
+      final String liveSegmentsLocation = ZKPaths.makePath(zkPaths.getLiveSegmentsPath(), me.getName());
+
+      loadQueueCache = new PathChildrenCache(
+          curator,
+          loadQueueLocation,
+          true,
+          true,
+          Execs.multiThreaded(config.getNumLoadingThreads(), "ZkCoordinator-%s")
+      );
+
+      try {
+        curator.newNamespaceAwareEnsurePath(loadQueueLocation).ensure(curator.getZookeeperClient());
+        curator.newNamespaceAwareEnsurePath(servedSegmentsLocation).ensure(curator.getZookeeperClient());
+        curator.newNamespaceAwareEnsurePath(liveSegmentsLocation).ensure(curator.getZookeeperClient());
+
+        loadLocalCache();
+
+        loadQueueCache.getListenable().addListener(
+            new PathChildrenCacheListener()
+            {
+              @Override
+              public void childEvent(CuratorFramework client, PathChildrenCacheEvent event) throws Exception
+              {
+                final ChildData child = event.getData();
+                switch (event.getType()) {
+                  case CHILD_ADDED:
+                    final String path = child.getPath();
+                    final DataSegmentChangeRequest request = jsonMapper.readValue(
+                        child.getData(), DataSegmentChangeRequest.class
+                    );
+
+                    log.info("New request[%s] with zNode[%s].", request.asString(), path);
+
+                    try {
+                      request.go(
+                          getDataSegmentChangeHandler(),
+                          new DataSegmentChangeCallback()
+                          {
+                            boolean hasRun = false;
+
+                            @Override
+                            public void execute()
+                            {
+                              try {
+                                if (!hasRun) {
+                                  curator.delete().guaranteed().forPath(path);
+                                  log.info("Completed request [%s]", request.asString());
+                                  hasRun = true;
+                                }
+                              }
+                              catch (Exception e) {
+                                try {
+                                  curator.delete().guaranteed().forPath(path);
+                                }
+                                catch (Exception e1) {
+                                  log.error(e1, "Failed to delete zNode[%s], but ignoring exception.", path);
+                                }
+                                log.error(e, "Exception while removing zNode[%s]", path);
+                                throw Throwables.propagate(e);
+                              }
+                            }
+                          }
+                      );
+                    }
+                    catch (Exception e) {
+                      try {
+                        curator.delete().guaranteed().forPath(path);
+                      }
+                      catch (Exception e1) {
+                        log.error(e1, "Failed to delete zNode[%s], but ignoring exception.", path);
+                      }
+
+                      log.makeAlert(e, "Segment load/unload: uncaught exception.")
+                         .addData("node", path)
+                         .addData("nodeProperties", request)
+                         .emit();
+                    }
+
+                    break;
+                  case CHILD_REMOVED:
+                    log.info("zNode[%s] was removed", event.getData().getPath());
+                    break;
+                  default:
+                    log.info("Ignoring event[%s]", event);
+                }
+              }
+            }
+        );
+        loadQueueCache.start();
+      }
+      catch (Exception e) {
+        Throwables.propagateIfPossible(e, IOException.class);
+        throw Throwables.propagate(e);
+      }
+
+      started = true;
+    }
+  }
+
+  @LifecycleStop
+  public void stop()
+  {
+    log.info("Stopping ZkCoordinator for [%s]", me);
+    synchronized (lock) {
+      if (!started) {
+        return;
+      }
+
+      try {
+        loadQueueCache.close();
+      }
+      catch (Exception e) {
+        throw Throwables.propagate(e);
+      }
+      finally {
+        loadQueueCache = null;
+        started = false;
+      }
+    }
+  }
+
+  public boolean isStarted()
+  {
+    return started;
+  }
+
   public void loadLocalCache()
   {
     final long start = System.currentTimeMillis();
@@ -129,7 +282,6 @@ public class ZkCoordinator extends BaseZkCoordinator
     );
   }
 
-  @Override
   public DataSegmentChangeHandler getDataSegmentChangeHandler()
   {
     return ZkCoordinator.this;
@@ -168,7 +320,7 @@ public class ZkCoordinator extends BaseZkCoordinator
   {
     try {
       log.info("Loading segment %s", segment.getIdentifier());
-      if(loadSegment(segment, callback)) {
+      if (loadSegment(segment, callback)) {
         try {
           announcer.announceSegment(segment);
         }
@@ -203,24 +355,34 @@ public class ZkCoordinator extends BaseZkCoordinator
       final CopyOnWriteArrayList<DataSegment> failedSegments = new CopyOnWriteArrayList<>();
       for (final DataSegment segment : segments) {
         loadingExecutor.submit(
-            new Runnable() {
+            new Runnable()
+            {
               @Override
-              public void run() {
+              public void run()
+              {
                 try {
-                  log.info("Loading segment[%d/%d][%s]", counter.getAndIncrement(), numSegments, segment.getIdentifier());
+                  log.info(
+                      "Loading segment[%d/%d][%s]",
+                      counter.getAndIncrement(),
+                      numSegments,
+                      segment.getIdentifier()
+                  );
                   final boolean loaded = loadSegment(segment, callback);
                   if (loaded) {
                     try {
                       backgroundSegmentAnnouncer.announceSegment(segment);
-                    } catch (InterruptedException e) {
+                    }
+                    catch (InterruptedException e) {
                       Thread.currentThread().interrupt();
                       throw new SegmentLoadingException(e, "Loading Interrupted");
                     }
                   }
-                } catch (SegmentLoadingException e) {
+                }
+                catch (SegmentLoadingException e) {
                   log.error(e, "[%s] failed to load", segment.getIdentifier());
                   failedSegments.add(segment);
-                } finally {
+                }
+                finally {
                   latch.countDown();
                 }
               }
@@ -228,14 +390,15 @@ public class ZkCoordinator extends BaseZkCoordinator
         );
       }
 
-      try{
+      try {
         latch.await();
 
-        if(failedSegments.size() > 0) {
+        if (failedSegments.size() > 0) {
           log.makeAlert("%,d errors seen while loading segments", failedSegments.size())
-              .addData("failedSegments", failedSegments);
+             .addData("failedSegments", failedSegments);
         }
-      } catch(InterruptedException e) {
+      }
+      catch (InterruptedException e) {
         Thread.currentThread().interrupt();
         log.makeAlert(e, "LoadingInterrupted");
       }
@@ -244,8 +407,8 @@ public class ZkCoordinator extends BaseZkCoordinator
     }
     catch (SegmentLoadingException e) {
       log.makeAlert(e, "Failed to load segments -- likely problem with announcing.")
-          .addData("numSegments", segments.size())
-          .emit();
+         .addData("numSegments", segments.size())
+         .emit();
     }
     finally {
       callback.execute();
@@ -298,7 +461,8 @@ public class ZkCoordinator extends BaseZkCoordinator
     }
   }
 
-  private static class BackgroundSegmentAnnouncer implements AutoCloseable {
+  private static class BackgroundSegmentAnnouncer implements AutoCloseable
+  {
     private static final EmittingLogger log = new EmittingLogger(BackgroundSegmentAnnouncer.class);
 
     private final int intervalMillis;
